# [wip] improved serializer performance