baseline

|                    Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
| --------------------------- |---------:|----------:|----------:|----------:|----------:|
|   Serialize_10000_Messages | 3.401 ms | 0.0663 ms | 0.0737 ms |  835.9375 |   3.36 MB |
| Deserialize_10000_Messages | 8.870 ms | 0.2068 ms | 0.2124 ms | 2828.1250 |  11.33 MB |

updated


 |                     Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
 |--------------------------- |---------:|----------:|----------:|----------:|----------:|
 |   Serialize_10000_Messages | 2.385 ms | 0.0444 ms | 0.0370 ms |  562.5000 |   2.25 MB |
 | Deserialize_10000_Messages | 7.616 ms | 0.1460 ms | 0.1562 ms | 2609.3750 |  10.45 MB |